Move enikshay case utils
