+++ custom/enikshay/exceptions.py
@@ -1,0 +1,6 @@
+class ENikshayException(Exception):
+    pass
+
+
+class ENikshayCaseNotFound(Exception):
+    pass
