+++ custom/enikshay/integrations/ninetyninedots/exceptions.py
@@ -1,3 +1,2 @@
 class AdherenceException(Exception):
-    def __init__(self, message):
-        self.message = message
+    pass
